feat(ResourceManager): wait until page has finished loading to continue fix(safe-browsing): check for API key on honeyclient startup refactor(safe-browsing): move into its own class fix(yara): wait until all files have been processed to exitt refactor(yara): extract code from callback hell refactor(yara): centralize error handling refactor(yara): move to be more in-line with the other managers feat(async-worker): EventEmitter paradigm for ops which require deferred callbacks
